+++ libraries/provider_firewall_iptables_ubuntu.rb
@@ -98,6 +98,8 @@ class Chef
         end
       end
 
+      restart_service = false
+
       rule_files = %w(iptables)
       rule_files << 'ip6tables' if ipv6_enabled?(new_resource)
 
@@ -120,12 +122,14 @@ class Chef
         iptables_file.run_action(:create)
 
         # if the file was changed, restart iptables
-        next unless iptables_file.updated_by_last_action?
+        restart_service = true if iptables_file.updated_by_last_action?
+      end
+
+      if restart_service
         service_affected = service 'netfilter-persistent' do
           action :nothing
         end
-
-        new_resource.notifies(:restart, service_affected, :delayed)
+        service_affected.run_action(:restart)
         new_resource.updated_by_last_action(true)
       end
     end